% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_preparation.R
\name{addEICs}
\alias{addEICs}
\title{Get EIC data on specified sequences}
\usage{
addEICs(dat, files, type = c("mzXML", "rds", "obj"), times, dir,
  verbose = TRUE)
}
\arguments{
\item{dat}{peptide sequences to extract, from makeSequences}

\item{files}{data files to read, or optionally, the objects themselves}

\item{type}{the type of data file, or R object}

\item{times}{the data collection times}

\item{dir}{the directory the files are in}

\item{verbose}{if verbose output desired}
}
\value{
this returns a list of lists, one for each protein, each with
a list for each sequence of that protein, each with
  \item{ID}{(a single character string)}
  \item{Sequence}{(a single character string)}
  \item{Times}{}
  \item{Elements}{}
  \item{data}{data frame with:
     MZ: the mz,
     RT: the retention times,
     Count: matrix with columns for channels and rows for retention times.}
}
\description{
Get EIC data on specified sequences. This extracts the raw eic values, not the profiled values in class.
}

