% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes8.R
\name{calcEnv}
\alias{calcEnv}
\title{Calculate natural abundance of isotopes over several elements.}
\usage{
calcEnv(..., abundance = NULL)
}
\arguments{
\item{abundance}{A list of the natural abundance for each element.
By default, it knows about C, H, O, N, and S.}

\item{\dots}{The number of each element of interest.}
}
\value{
A numeric vector with the natural abundance of each unit mass extra,
starting with zero extra.
}
\description{
For a set of elements, compute the total natural abundance.
}
\details{
It computes the natural abundance for each element, and convolves
them together to get the total natural abundance.
}
\examples{
# natural abundance for 1 nitrogen element
calcEnv(N=1)
 
# combined natural abundance for 2 nitrogen elements
calcEnv(N=2)
 
# combined natural abundance for a new element Z
calcEnv(Z=1, abundance=list(Z=c(0.7, 0.2, 0.1)))
 
# combined natural abundance for Z and N
calcEnv(N=1, Z=1, abundance=list(Z=c(0.7, 0.2, 0.1)))
}

