% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{getPars}
\alias{getPars}
\title{Get output from a list of completed fits}
\usage{
getPars(seqs, score = score.visual)
}
\arguments{
\item{seqs}{A list of completed fits. Each element should be a list with three elements:\itemize{
  \item{seq}{the original sequence information}
  \item{relAb}{the relative abundance}
  \item{fit}{the fit}
}}

\item{score}{the function to compute the score with}
}
\value{
a data frame with the output
}
\description{
Get output from a list of completed fits.
}
