% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{isocounts}
\alias{isocounts}
\title{Sample Count Data}
\format{A data frame with 1224 observations on the following 4 variables.
 \describe{
   \item{\code{TimePoint}}{Day the observation was made on}
   \item{\code{Channel}}{Channel the observation was on (that is, how
     many ``heavy'' elements there were)}
   \item{\code{RT}}{Retention time the observation was made at}
   \item{\code{Count}}{Number of items observed}
 }}
\description{
A sample of the kind of count data used by this package.
}
\details{
Count data like this can be converted to relative abundances using relAbFromCounts.
}
\examples{
data(isocounts)
head(isocounts)
counts <- within(isocounts,{
  Channel <- factor(Channel)
  TimePoint <- factor(TimePoint)
  logCount <- log10(Count)
})
coplot(logCount~RT|Channel*TimePoint, data=counts, show.given=FALSE)
}
\keyword{datasets}
