% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{isodata}
\alias{isodata}
\title{Sample Peptide/Isotope Data}
\format{A data frame with 68 observations on the following 3 variables.
 \describe{
   \item{\code{TimePoint}}{The day the observation was made}
   \item{\code{Channel}}{Number of ``heavy'' elements}
   \item{\code{RelAb}}{The relative abundance of that count on that
     day, compared to the other counts on that day}
 }}
\description{
A sample of the kind of data used by this package.
}
\examples{
data(isodata)
head(isodata)
coplot(RelAb~Channel|factor(TimePoint), data=isodata)
pepfit(TimePoint, RelAb, Channel, data=isodata, Elements=list(N=12,C=45,H=73,O=15))
}
\keyword{datasets}

