% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{isoincorp}
\alias{isoincorp}
\title{Sample Peptide/Isotope Data}
\format{A data frame with 73 observations on the following 3 variables.
 \describe{
   \item{\code{TimePoint}}{The day the observation was made}
   \item{\code{Channel}}{Number of ``heavy'' elements}
   \item{\code{RelAb}}{The relative abundance of that count on that
     day, compared to the other counts on that day}
 }}
\description{
A sample of the incorporation data used by this package.
}
\examples{
data(isoincorp)
head(isoincorp)
coplot(RelAb~Channel|factor(TimePoint), data=isoincorp)
pepfit(TimePoint, RelAb, Channel, data=isoincorp,
Elements=list(N=12,C=45,H=82,O=12), type="incorporation")
}
\keyword{datasets}

