% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/protein_preparation.R
\name{makeSequence}
\alias{makeSequence}
\title{Extract peptide sequence information for a single sequence from a data file}
\usage{
makeSequence(x, pars, aa.table = aaTable)
}
\arguments{
\item{x}{the data to get the information from}

\item{pars}{a list with parameter information, as passed from makeSequences}

\item{aa.table}{the table to take element information from}
}
\value{
list with ID, Sequence, N, C, Channels, a list of data frames
  (one for each TimePoint) with dataframe par (RT, MZ) with
  possible multiple rows, if multiple found, and matriz MZ
  with MZ for each channel (as columns) for each row of par

if error, has "error" term with description of error,
  then doesn't have Times
}
\description{
Extract peptide sequence information for a single sequence from a data file
}
\details{
counts how many elements are present in the amino acid sequence
then builds out the channels based on the isotope under study
}

