% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{makeSequenceHTML}
\alias{makeSequenceHTML}
\title{Make HTML output for a list of fits}
\usage{
makeSequenceHTML(seqs, dir, file, subdir = paste0(file, "-seqs"))
}
\arguments{
\item{seqs}{A list of completed fits. Each element should be a list with three elements: \itemize{
  \item{seq}{the original sequence information}
  \item{relAb}{the relative abundance}
  \item{fit}{the fit}
}}

\item{dir}{the directory the output should be in}

\item{file}{the main file name (without the .html added)}

\item{subdir}{the subdirectory the peptide sequence results should go in}
}
\value{
information on the individual pages, invisibly
}
\description{
Make HTML output for a list of fits
}
