% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_preparation.R, R/run.R
\name{makeSequences}
\alias{makeSequences}
\alias{getData}
\title{Extract peptide sequence information from a data file}
\usage{
makeSequences(d, isotope, confidence.level, time.zero, time.unit, mz.tol,
  rt.tol, extra.channels, aa.table = aaTable, verbose = TRUE)

getData(seq)
}
\arguments{
\item{d}{the data file, from reading the Scaffold or ProteinPilot files}

\item{isotope}{the desired isotope}

\item{confidence.level}{the desired peptide identification confidence value}

\item{time.zero}{the time to treat as time "zero"}

\item{time.unit}{the name of the time unit, for display}

\item{mz.tol}{the m/z window}

\item{rt.tol}{the retention time window}

\item{extra.channels}{how many extra channels to include}

\item{aa.table}{the table to take element information from}

\item{verbose}{if verbose output desired}

\item{seq}{a sequence}
}
\value{
a list of peptide sequences, as from makeSequence
}
\description{
Extract peptide sequence information from a data file
}
