% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicore.R
\name{mclapply.progress}
\alias{mclapply.progress}
\title{lapply over multiple cores with progress bar}
\usage{
mclapply.progress(x, FUN, ..., mc.preschedule = TRUE, mc.set.seed = TRUE,
  mc.silent = FALSE, mc.cores = getOption("mc.cores", 1L),
  mc.cleanup = TRUE, mc.allow.recursive = TRUE)
}
\arguments{
\item{x}{the elements to loop over}

\item{FUN}{the function to apply}

\item{mc.preschedule}{passed to mclapply}

\item{mc.set.seed}{passed to mclapply}

\item{mc.silent}{passed to mclapply}

\item{mc.cores}{passed to mclapply}

\item{mc.cleanup}{passed to mclapply}

\item{mc.allow.recursive}{passed to mclapply}

\item{\dots}{parameters passed to FUN}
}
\value{
the result from mclapply(x, FUN, ...)
}
\description{
lapply over multiple cores with progress bar
}

