% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsetup.R
\name{parmatrix.byix}
\alias{parmatrix.byix}
\title{Create a function that will set up a parmatrix}
\usage{
parmatrix.byix(ix, dpar)
}
\arguments{
\item{ix}{A matrix that defines which value from the simple vector go
in that location.  If a default value is needed instead, the value should be
set to NA.}

\item{dpar}{A matrix that defines the default values for locations in the ix
matrix set to NA.  Values in other locations are ignored.}
}
\value{
A function that takes a simple vector of parameters and creates a
 matrix, repeating parameters and substituting default values as required.
}
\description{
Creates a function that takes a simple vector of parameters and creates a
  matrix, repeating parameters and substituting default values as required.
}
\details{
Suppose a function takes a matrix of predictors, and you want to
 optimize those predictors, but with some of them always set equal to
 each other and others always set to some default value.  Since the optimize function
 uses not the full matrix as input and output but instead, a simple
 vector of the parameters that are allowed to change.  This function
 creates a function that will convert this simple vector to the full
 matrix, repeating values as necessary and filling in default values.
}
\examples{
(par.index <- matrix(c(1,1,2,NA), nrow=2))
 
(par.default <- matrix(c(NA,NA,NA,100), nrow=2))
 
par.function <- parmatrix.byix(par.index, par.default)
par.function(c(1.1, 1.2))
par.function(c(21, 22))
}

