% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pep8.R
\name{parmatrix}
\alias{parmatrix}
\title{Gets the fitted parameters from a pepfit object}
\usage{
parmatrix(pepfit, default = TRUE, as.df = FALSE)
}
\arguments{
\item{pepfit}{A pepfit object}

\item{default}{Whether or not parameters given as defaults should be included or not}

\item{as.df}{Whether it should return a data frame or a matrix}
}
\value{
The fitted parameters from the pepfit object, in the desired format
}
\description{
Gets the fitted parameters from a pepfit object
}

