% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lackfit.R
\name{pepcompare}
\alias{pepcompare}
\alias{print.pepcompare}
\title{Compare two fitted models}
\usage{
pepcompare(pepfit1, pepfit2, sig.digits = 5)

\method{print}{pepcompare}(x, sig.digits = x$sig.digits, ...)
}
\arguments{
\item{pepfit1}{The first fitted model}

\item{pepfit2}{The second fitted model}

\item{sig.digits}{Number of significant digits to report.}

\item{x}{A pepcompare object}

\item{\dots}{Ignored, here for consistency with generic print function}
}
\value{
A list with class "pepcompare" and the following elements:
 \item{deviance}{A data frame with deviance and degrees of freedom for
   the two models, and the difference between them.}
 \item{N}{The input value "N"}
 \item{p.value}{The p value corresponding to the input value "N"}
 \item{level}{The input value "level"}
 \item{N.value}{The sample size corresponding to the input value
   "level"}
 \item{sig.digits}{Number of significant digits to report.}
}
\description{
Compare two models, fitted using different parameter options.
}
\details{
The models are compared by computing the change in deviance, but the
sample size is unknown, so a p-value cannot be computed without
  assuming a given sample size.
 
  Also, this assumes the probability distribution of deviance is
  chi-squared; this can be checked with the pepsim function.
}
\examples{
  data(isodata)
 a0 <- pepfit(TimePoint, RelAb, Channel, data=isodata, Elements = list(N = 12,
    C = 45, H = 73, O = 15))
 a2 <- pepfit(TimePoint, RelAb, Channel, data=isodata, Elements = list(N = 12,
    C = 45, H = 73, O = 15), M = "one")
 pepcompare(a0, a2)
}

