% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/curve.R, R/run.R
\name{pepcurve}
\alias{getCurve}
\alias{makeCurve}
\alias{pepcurve}
\alias{pepcurve.default}
\alias{pepcurve.pepfit}
\title{Fit an exponential curve to fitted pi, r, and alpha values}
\usage{
pepcurve(x, ...)

\method{pepcurve}{pepfit}(x, nab = TRUE, fillNA = FALSE, ...)

\method{pepcurve}{default}(x, nab = NA, fillNA = FALSE, intercept = c(pi =
  NA, alpha = NA, r = NA), ...)

makeCurve(fit, plim = NA, isotope.method = c("other", "incorporation"), ...)

getCurve(curve)
}
\arguments{
\item{x}{A pepfit object, or a data frame with columns TimePoint, and at
least one of pi, r, and alpha}

\item{nab}{for pepcurve.default, sets the natural abundance for use in either
intercept or the limit of pi; NA finds the best fit instead
for pepcurve.pepfit, can be numeric, or TRUE to use the known natural
abundance from the pepfit, or FALSE to find the best fit}

\item{fillNA}{should output have NA?}

\item{intercept}{set intercept values for pi, alpha, or r}

\item{fit}{fit from makeFit}

\item{plim}{sent to pepcurve as nab parameter}

\item{isotope.method}{used to set up intercepts correctly for incorporation experiments}

\item{curve}{a fitted curve}

\item{\dots}{used to pass nab to the appropriate function}
}
\value{
A list of class "pepcurve".

 The first element has four values. The first three are from the pi fit:
the fitted value at day 0, the lower limit, and the halflife.  The last
is the halflife from the r fit, for which the fitted value at day 0 is
set to 1 and the lower limit is set to 0.

Remaining elements contain the pi and r values from the
pepfit model and the fitted curve objects for pi and r, I think.  This
should be checked to make sure it's still true.
}
\description{
Fit an exponential curve to fitted pi, r, and alpha values
}
\examples{
data(isodata)
a0 <- pepfit(TimePoint, RelAb, Channel, data=isodata, Elements = list(N = 12,
     C = 45, H = 73, O = 15))

c0 <- pepcurve(a0)
c0
plot(c0)
}

