% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lackfit.R
\name{pepfittest}
\alias{pepfittest}
\alias{print.pepfittest}
\title{Test a pepfit}
\usage{
pepfittest(object)

\method{print}{pepfittest}(x, sig.digits = 5, ...)
}
\arguments{
\item{object}{a pepfit object}

\item{x}{the pepfittest object}

\item{sig.digits}{desired number of significant digits}

\item{\dots}{to match the generic print function}
}
\value{
A list of class pepfittest, with elements
 \item{deviance.fit}{deviance of the fitted model}
 \item{deviance.null}{deviance of the null model}
 \item{X2}{Pearson X^2 of the fitted model}
 \item{nobs}{number of observations in data set}
 \item{df.fit}{degrees of freedom for the fitted model}
 \item{df.null}{degrees of freedom for the null model}
 \item{sigma2}{the quasi-multinomial scaling paramter}
 \item{R2}{the Nagelkerke R^2}
 \item{data}{observed, expected, and deviance values for each observation}
}
\description{
Test a pepfit.  This is usually not run alone, as it is included in the summary of a pepfit.
}
\examples{
data(isodata)
a0 <- pepfit(data=isodata, Elements = list(N = 12,
     C = 45, H = 73, O = 15))
pepfittest(a0)
}
\seealso{
\code{\link{pepfit}}
}

