% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lackfit.R
\name{pepsim}
\alias{pepsim}
\alias{plot.pepsim}
\alias{print.pepsim}
\title{Create a set of simulated data}
\usage{
pepsim(object, Nobs = 1000, Nrep = 10)

\method{print}{pepsim}(x, ...)

\method{plot}{pepsim}(x, n = 30, ...)
}
\arguments{
\item{object}{a pepfit object}

\item{Nobs}{the sample size to be used for the simulation}

\item{Nrep}{how many simulations to run}

\item{x}{a pepsim object}

\item{n}{how many simulations to plot}

\item{\dots}{currently ignored}
}
\value{
A list of class pepsim containing:
 \item{sim}{a data frame of simulated relative abundances for each day
   and count value}
 \item{Nobs}{Nobs, as input}
 \item{Nrep}{Nrep, as input}
 \item{parent}{the original pepfit object}
}
\description{
Create a set of simulated data, to check if the change in deviance
between two models is really chi-squared.
}
\examples{
data(isodata)
a0 <- pepfit(data=isodata, Elements = list(N = 12,
    C = 45, H = 73, O = 15))
a0.sim <- pepsim(a0, Nrep = 50)
plot(plot(a0.sim))
}

