% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lackfit.R
\name{pepsimtest}
\alias{pepsimtest}
\alias{plot.pepsimtest}
\alias{print.pepsimtest}
\title{Test a set of simulated data}
\usage{
pepsimtest(object, ask = TRUE, n = object$Nrep, model = object$parent)

\method{print}{pepsimtest}(x, ...)

\method{plot}{pepsimtest}(x, ylim, ...)
}
\arguments{
\item{object}{a pepsim object}

\item{ask}{Since this may take a while, it will ask for permission before
starting.  Disable with ask=FALSE.}

\item{n}{The number of simulations to fit models to.}

\item{model}{A pepfit object with the model that you want to fit the simulated data to.}

\item{x}{the pepsimtest object}

\item{ylim}{the vertical limits of the plot}

\item{\dots}{to match the generic print function}
}
\value{
A list of class "pepsimtest" containing:
 \item{sim.statistic}{The deviances from the fitted models for each
   simulation}
 \item{obs.test}{The test of the original fitted model}
 \item{sim}{The original pepsim object}
}
\description{
Fit a separate model to each simulation and find the deviance, to see if
the distribution of the deviance is really chi-squared.
}
\examples{
\dontrun{
data(isodata)
a0 <- pepfit(data=isodata, Elements = list(N = 12,
    C = 45, H = 73, O = 15)))
a0.sim <- pepsim(a0, Nrep = 50)
a0.test.sim <- pepsimtest(a0.sim, n = 10, ask = FALSE)
a0.test.sim
plot(a0.test.sim)
}
}
\seealso{
\code{\link{pepsim}}
}

