% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_preparation.R
\name{rawToRDS}
\alias{rawToRDS}
\title{Convert raw mzXML files to Rdata files}
\usage{
rawToRDS(dir, files, Rfiles)
}
\arguments{
\item{dir}{directory that files are in}

\item{files}{names of mzXML files}

\item{Rfiles}{name of Rdata files}
}
\value{
names of Rdata files (invisibly)
}
\description{
Convert raw mzXML files to Rdata files, and save in the same directory.
}
