% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_preparation.R
\name{readScaffoldFile}
\alias{readScaffoldFile}
\title{Read Scaffold file}
\usage{
readScaffoldFile(file, times, dir, verbose = TRUE)
}
\arguments{
\item{file}{name of Scaffold file}

\item{times}{data collection times, as a vector}

\item{dir}{directory that file is in}

\item{verbose}{if verbose output desired}
}
\value{
data frame with: ID, Sequence, RT, MW, Z, TimePoint, Conf
}
\description{
Read Scaffold file
}

