% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions2.R, R/run.R
\name{relAbForTimes}
\alias{relAbForTimes}
\alias{regressionPlot}
\alias{plot.RelAbTimes}
\alias{makeRelAb}
\title{Relative Abundance from Counts, for multiple time points}
\usage{
relAbForTimes(Count, Channel, RT, TimePoint, data, nboot = 0, ...)

regressionPlot(x, ...)

\method{plot}{RelAbTimes}(x, ..., type = c("relAb", "regression"))

makeRelAb(seq, regression.model, nboot = 0)
}
\arguments{
\item{Count}{Observed counts}

\item{Channel}{Channels corresponding to the observed counts}

\item{RT}{Retention times corresponding to the observed counts}

\item{TimePoint}{Time points corresponding to the observed counts}

\item{data}{An optional data frame to take the preceding variables from}

\item{nboot}{Experimental, for bootstrapping}

\item{\dots}{Additional parameters to be passed}

\item{x}{A regRelAbTimes or logRelAbTimes object to be plotted}

\item{type}{desired type of plot}

\item{seq}{the sequence to get the relative abundance for}

\item{regression.model}{passed as method to relAbFromCounts}
}
\value{
A list is returned with the following elements.  It is of class
 "regRelAbTimes" if a regression method is used and "logRelAbtimes" if a log
 model is used.
 \item{data.long}{The long data frames from each fit, bound together
   into one data frame.}
 \item{norm_channel}{The chosen baseline channel for each timepoint}
 \item{coefs}{For the regression model, a list with the fitted coefficients at each
 time points.}
 \item{relab}{The fitted relative abundance at each channel and timepoint}
 \item{proportion}{The fitted relative abundance, converted to a
   proportion}
 \item{method}{The method used}
 \item{threshold}{The threshold used}
}
\description{
Gets relative abundance of each channel from counts taken at several
 retention times, and over multiple time points
}
\details{
The data from each time point is fit separately, using
 relAbFromCounts.  Plots are created using the lattice library.
}
\examples{
data(isocounts)
}
\seealso{
\code{\link{relAbFromCounts}}
}
