% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proportions.R
\name{relAbFromCounts}
\alias{relAbFromCounts}
\title{Relative Abundance from Counts}
\usage{
relAbFromCounts(Count, Channel, RT, data, TimePoint, method = c("lm", "rlm",
  "lqs", "rq", "sum", "log"), norm_channel, threshold, cutoff, nboot = 0)
}
\arguments{
\item{Count}{Observed counts}

\item{Channel}{Channels corresponding to the observed counts}

\item{RT}{Retention times corresponding to the observed counts}

\item{data}{An optional data frame to take the preceding variables from.}

\item{TimePoint}{The time point that this data is from, to be optionally added to the
output data frames.}

\item{method}{Options are "lm", "rlm", "lqs", "rq", "sum", and "log".}

\item{norm_channel}{The channel to use as baseline for the regression methods, in the
output of the relative abundance, and when computing the threshold.
By default, the channel with the highest total counts is used.}

\item{threshold}{A proportion used to remove values corresponding to retention times
for which the value of the norm_channel is less than this threshold
times the max of the norm_channel.}

\item{cutoff}{All counts less than or equal to this are set to missing.}

\item{nboot}{Experimental, for bootstrapping}
}
\value{
A list is returned with the following elements.  It is of class
 "regRelAb" if a regression method is used and "logRelAb" if a log
 model is used.
 \item{data.wide}{A table of counts, by Channel, and RT}
 \item{data.long}{The counts in long data frame format}
 \item{data}{The fitted relative abundance and proportions at each channel.}
 \item{norm_channel}{The chosen baseline channel}
 \item{coefs}{For the regression model, the fitted coefficients}
 \item{method}{The method used}
 \item{threshold}{The threshold used}
}
\description{
Gets relative abundance of each channel from counts taken at several retention times
}
\details{
The "lm", "rlm", "lqs", and "rq" methods plots the counts for each channel against the
baseline channel and computes a regression; the slope is the relative
abundance.

The "sum" method simply adds the counts and gets the proportion in
each channel.

The "log" method takes the logs of the counts and computes a two-way
ANOVA on RT and Channel; the coefficients for the channels are then
converted back to relative abundances.

Plots are created using the lattice library.
}
\examples{
data(isocounts)
}

