% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep8.R
\name{summary.pepfit}
\alias{summary.pepfit}
\title{Summarize and print results from a pepfit}
\usage{
\method{summary}{pepfit}(object, digits = 3, ...)
}
\arguments{
\item{object}{A pepfit object}

\item{digits}{How many digits to print}

\item{\dots}{Additional arguments passed to \code{\link{pepfittest}}}
}
\value{
list of class summary.pepfit with elements
 \item{par}{Fitted parameters}
 \item{hessian}{Fitted hessian matrix, if in original pepfit}
 \item{time}{Elapsed time}
 \item{test}{the results from \code{\link{pepfittest}}}
}
\description{
Summarize and print results from a pepfit
}
\examples{
data(isodata)
a <- pepfit(data=isodata,
Elements=list(N=12,C=45,H=73,O=15))
summary(a)
}
\seealso{
\code{\link{pepfittest}}
}
