% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lackfit.R
\name{pepfittest}
\alias{pepfittest}
\alias{print.pepfittest}
\title{Test a pepfit}
\usage{
pepfittest(object, level = 0.05, N = 1000)

\method{print}{pepfittest}(x, sig.digits = 5, ...)
}
\arguments{
\item{object}{a pepfit object}

\item{level}{The significance level to compute the associated sample size for}

\item{N}{The sample size to compute the associated p value for}

\item{x}{the pepfittest object}

\item{sig.digits}{desired number of significant digits}

\item{\dots}{to match the generic print function}
}
\value{
A list of class pepfittest, with elements
 \item{statistic}{change in deviance from the full model to the
   fitted model}
 \item{df}{change in degrees of freedom}
 \item{deviance.fit}{deviance of the fitted model}
 \item{deviance.full}{deviance of the full model}
 \item{deviance.byday}{deviance for each day, separately}
 \item{df.fit}{degrees of freedom for the fitted model}
 \item{df.full}{degrees of freedom for the full model}
 \item{p.value}{p.value associated with the desired sample size}
 \item{N}{the sample size used to compute the p.value}
 \item{N.value}{the sample size associated with the desired significance level}
 \item{level}{the significance level used to compute N.value}
 \item{data}{observed, expected, and deviance values for each observation}
}
\description{
Test a pepfit.  This is usually not run alone, as it is included in the summary of a pepfit.
}
\examples{
data(isodata)
a0 <- pepfit(data=isodata, Elements = list(N = 12,
     C = 45, H = 73, O = 15))
pepfittest(a0)
}
\seealso{
\code{\link{pepfit}}
}

